requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.1.3
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.18
